Added endpoint for creating certificate requests

+++ Tim17_Back/src/main/java/com/ib/Tim17_Back/validations/UserRequestValidation.java
@@ -55,7 +55,7 @@ public class UserRequestValidation {
         return givenId.intValue() == id;
     }
 
-    public Long getUserId(Map<String, String> headers) {
+    public Integer getUserId(Map<String, String> headers) {
         String token = headers.get("x-auth-token");
         String refreshToken = headers.get("refreshtoken");
         if (isTokenExpired(token))
